+++ src/main/java/de/maxhenkel/peek/config/PeekConfig.java
@@ -3,7 +3,7 @@ package de.maxhenkel.peek.config;
 import de.maxhenkel.configbuilder.ConfigBuilder;
 import de.maxhenkel.configbuilder.ConfigEntry;
 
-public class ClientConfig {
+public class PeekConfig {
 
     public final ConfigEntry<Boolean> showEmptyContainers;
 
@@ -23,8 +23,9 @@ public class ClientConfig {
     public final ConfigEntry<Boolean> useShulkerBoxDataStrings;
     public final ConfigEntry<Boolean> hideShulkerBoxDataStrings;
     public final ConfigEntry<Integer> shulkerBoxItemHintCustomModelData;
+    public final ConfigEntry<Boolean> sendShulkerBoxDataToClient;
 
-    public ClientConfig(ConfigBuilder builder) {
+    public PeekConfig(ConfigBuilder builder) {
         showEmptyContainers = builder.booleanEntry("show_empty_containers", false);
 
         peekShulkerBoxes = builder.booleanEntry("peek_shulker_boxes", true);
@@ -42,7 +43,8 @@ public class ClientConfig {
         showShulkerBoxBlockHint = builder.booleanEntry("show_shulker_box_block_hint", true);
         useShulkerBoxDataStrings = builder.booleanEntry("use_shulker_box_data_strings", false);
         hideShulkerBoxDataStrings = builder.booleanEntry("hide_shulker_box_data_strings", true);
-        shulkerBoxItemHintCustomModelData = builder.integerEntry("shulher_box_item_hint_custom_model_data", 1702, 0, Integer.MAX_VALUE);
+        shulkerBoxItemHintCustomModelData = builder.integerEntry("shulker_box_item_hint_custom_model_data", 1702, 0, Integer.MAX_VALUE);
+        sendShulkerBoxDataToClient = builder.booleanEntry("send_shulker_box_data_to_client", true);
     }
 
 }